requirements.txt
pymysql~=1.0.2
Django>2
requests>=2.12
loguru~=0.6.0
djangorestframework~=3.14.0
PyYAML~=6.0
